git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@3153 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ src/workbench/sql/wbcommands/WbOraShow.java
@@ -350,13 +350,13 @@ public class WbOraShow
 	private long roundToKb(long input)
 	{
 		if (input < ONE_KB) return input;
-		return (long)(input / ONE_KB);
+		return input / ONE_KB;
 	}
 
 	private long roundToMb(long input)
 	{
 		if (input < ONE_MB) return input;
-		return (long)(input / ONE_MB);
+		return input / ONE_MB;
 	}
 
 }
